Adds the /onpush hook into the server.
Implemented via pulling from GitHub, then reading all data from the
filesystem, and subsequently write-locking the data store, and updating
it. Locking the data store for as minimal a time as possible was the
primary intent; this should cause minimal downtime for users (since the
data store should only be write locked for a short time).