Replace start_idx and end_idx with index_in.
Start_idx and end_idx are conceptually the same function; their
difference in JS historically was that start_idx returned 0 when there
were 0 or 1 elements in the passed array while end_idx returned the last
element's index.

This distinction is relatively pointless: when searching for a date in
the array, we want to find the closest date to it, returning that date's
index; it makes no difference whether we are searching for a start date
or an end date.