Add 'Updated as of' text to the bottom of each page
The date is computed by taking the last chronological date from the
loaded data.

No effort is made to link to where the data is normally loaded from, or
to note who to notify if the reader believes there to be a problem.
While doing so is now simple, I don't believe it necessary to include in
the as-of text. This should instead be on something akin to an about
page.