Make cells on summary page link to time graph
Only the time graph is currently supported, since most of the time, it
is the one the user cares about.

The total column is not supported since it is rare that someone will
care about improvement over the past 13 weeks, but it should be
relatively simple to add it if requested; though the changes will not be
trivial.

The implementation currently queries the info endpoint for the phase
list to pass to the graphs page, but the ideal way to do this would be a
meta "all_phases" phase name we could pass instead.