Correct expected results to the expected values
Serde's float deserialization/serialization currently loses precision,
which lead to problems where different numbers of roundtrips through
Serde would cause differences in output. To fix the tests, we now
compare two `serde_json::Value`s that are the result of an equal number
of roundtrips.

A commented line which writes (new) expected results to the files is
added. The intended use case would be that whenever tests change, it is
uncommented, tests are re-run, and [git] diff is used to compare the
results to figure out what changed.

The difference and term crates are removed from the dependencies since
we no longer greatly benefit from displaying the diffs when tests fail.
This does hurt CI results, since it is now slightly harder to tell why
tests failed; but since CI's primary intent is not to give the why but
to give a boolean yes/no answer, this is considered okay by the author.
Specifically, the raw server response is still displayed, but the diff
between the expected result and the actual result is not.