Display only the current series in graph tooltips
It is hard to tell which data point is the current one when many series
are displayed, so we should only display one series in the graph
tooltip. Relative comparisons more so than exact comparisons are
generally made on the graph page, and exact data often varies from
run-to-run even without changes to the compiler, so this change should
not hurt usability overall, and in fact should improve it.

Potential future work could involve finding a way to query the previous
and next series in a given "column" on the graph, and display the series
data for those two data points.