Transition to a uniform benchmark approach.
All benchmarks get the same basic states:

Clean (no incr), Incr initial, Incr Clean, and then some number of
patches.

No benchmark can change this for itself at the moment; but it shouldn't
need to -- any more states should be added to all benchmarks, ideally.

Note: This commit does not update the benchmarks or frontend, only the
collector. This means that it does not compile (site is broken) and the
benchmarks won't work well.