Use -is-final-crate to indicate the final crate.
Current -Ztime-passes is used, but it's misleading because it's treated
in a special way by fake-rustc. In contrast, -is-final-crate is not a
valid rust argument, and so it's more obvious that something unusual is
going on.

The patch also tweaks the exec() function in rustc-fake to be clearer --
putting cmd.exec() inside panic!() is rather non-obvious.