Fix the 'bench_local' subcommand.
Because it is currently pretty much unusable.

This patch changes its invocation to this:

  collector bench_local <ID> --cargo <CARGO> --rustc <RUSTC>

<ID> is an identifier used in the output file's name and contents, in the same
places that the commit SHA is used when benchmarking a particular commit.
<CARGO> and <RUSTC> are mandatory options specifying local versions of Cargo
and rustc.

Output is now produced in much the same way as the other subcommands,
with the exception that the directory named by --output-repo doesn't
have to be an actual rustc-timings repo; any directory will do. `site`
can be used to view the data in the output directory.

Details of note:

- The patch fixes up the README.md and the usage docs a bit.

- Sysroot is no longer used in bench_commit() or execute.rs. Instead we pass
  `rustc_path`, `cargo_path`, and (for bench_commit() only) `triple`, because
  that's all that is needed.

- Sysroot::command() is no longer used by CargoProcess::base_command().
  It was only contributing a small amount of functionality (e.g. the
  PATH and CARGO env vars), which CargoProcess::base_command() can do
  itself.

- Repo::open() gains an additional parameter that tells it not to complain if
  the 'retries' file is missing; this is necessary now that ordinary
  directories can be used for output by 'bench_local'.