Tweak specification of incremental builds.
This patch makes `CargoProcess::incremental` set via a setter() func, like
`nll` and `perf`, instead of being a mandatory argument to
`mk_cargo_process()`.

This removes the need for `incremental` to be a newtype, so the patch
converts it to a plain bool.

The patch also moves a couple of things around so that `incremental`,
`nll` and `perf` are always together.