Remove the USE_PERF env var.
It's redundant with respect to -is-final-crate -- there is no case in
which we want one enabled but not the other. It's also not necessary for
the `cargo pkgid` invocations.

This change requires only setting -is-final-crate when perf is required,
rather than setting it unconditionally.

The change also requires changing run_rustc() to handle the case where
perf output is not present. (This change was necessary because
rustc-fake was buggy -- it always ran perf, even if USE_PERF was set to
0.)

In rustc-fake, we can now combine the three separate conditions in
main() into one, simplifying it.

Finally, the patch also renames -is-final-crate as
--wrap-rustc-with, and gives it an argument -- the only valid argument
is currently "perf-stat", though the next patch will change that. This
new name better reflects how it is used.