Remove assumption that cases for benchmarks do not change over time
Specifically, when a case is present in column A but not present in
column B we will skip over it (printing a - for B) but maintain
consistency in column order. The code currently presumes that cases
won't be reordered over time, but that feels like a much more valid
assumption, and could be corrected later on if necessary.