Mark measurements that have high variation.
A couple of the rustc-perf benchmarks have high variation for particular
runs, which can lead to misleading results.

This patch adds a mechanism for marking such results. Here's an example of what the output looks like.
```
              style-servo          avg: 3.8%?         min: -0.4%?  max: 19.3%?
                    clean  327,201,649,786.00  327,079,069,169.00        -0.0%
     baseline incremental  580,869,925,861.00  581,642,436,771.00         0.1%
        clean incremental  107,184,307,330.00  127,902,127,040.00      19.3%??
     patched incremental:
        debugging println  336,986,389,937.00  335,518,589,160.00        -0.4%
     patched incremental:
 b9b3e592dd cherry picked  470,140,412,082.00  470,324,586,916.00         0.0%
```
Unreliable measurements are marked with "??". Values derived from unreliable
measurements are marked with "?". Furthermore, all marked measurements have a
"title" element explaining how they are unreliable, so you get a tool-tip if
you hover over them in a desktop browser.

The particular runs marked as unreliable are both of the "patched incremental"
runs for coercions, and the "clean incremental" run for style-servo. These are
based on my experience.