Reduce uses of `RustcFeatures::is_stable`.
`RustcFeatures::is_stable` is currently one of the things controlling whether
or not NLL runs should occur. But `RunKind` also controls whether NLL runs
should occur, and is more general. So this patch removes the use of `is_stable`
to control NLL runs. This removes the need for `RustcFeatures` within
`CargoProcess`, and the `RunKind` selection logic is now entirely within
`main.rs`.