Merge `Benchmark::measure()` and `Benchmark::profile()`.
These two functions have a lot of overlap. This patch introduces the
`Processor` trait, which abstracts away the differences between the two
functions. There are two implementations of the new trait,
`MeasureProcessor` (which implements the `measure()`-specific
functionality) and `ProfileProcessor` (which implements the
`profile()`-specific functionality).

As a result `profile()` is no longer necessary, and `measure()` is
passed a `Processor` trait object.