Add "-debug" suffix.
Currently check builds have a "-check" suffix, opt builds have a "-opt"
suffix, and debug builds have no suffix. This patch adds a "-debug"
suffix to debug builds, which makes it clearer what they represent, and
the presentation more consistent overall.

As a result, on the "graphs" page the order from left to right is now
"-check", "-debug", "-opt", which is nice because it now goes from the
smallest execution time to the largest.