Rename existing data series, and add "clean" data series.
The renaming of the data series makes the dashboard use the same
terminology as the other views.

The addition of the "clean" data series provides continuity between the
pre- and post-incremental compilation eras, and it's also an interesting
measurement in its own right (and one that is present in the other
views).