Display newly added entries on comparison display
When an entry is absent in commit A but present in commit B we used to
not show it, but this is wrong -- we should instead show this as `-` to
some number. This is now the case.

The percent_chg function is left the same as before. This may not be
quite the desired behavior as it hides the presence/absence of data;
however, that change is out of scope for this commit.