Fix memory "leak"
It seems that most allocators don't cope well with our strategy for
refreshing the data from disk.

The refresh currently happens within a threadpool, which means that
we're creating quite a few threads (equal to number of vCPUs) on some
machines.

However, for most machines, even a few threads are sufficient to cause
problems with the system allocator. The arenas it creates do not appear
to be reduced in size when the memory is freed, so we end up with
multiple ~5 GB sized arenas (the current size of the data in memory).

This pulls in jemalloc and enables background threads. Based on local
testing, jemalloc itself does little to deal with this problem: the
memory "leak" still occurs. Background threads solve the problem by
releasing unused memory.