Add `cranelift-codegen` benchmark.
This commit imports Cranelift (revision
84567979ba18ad613d0066c0d8129393c9098a9c) into rustc-perf, and uses its
`cranelift-codegen` crate, which is by far its biggest crate, as the
benchmark.

The commit imports all of Cranelift even though only one of its crates
is benchmarked. This is consistent with how `style-servo` and
`script-servo` were done.

On my machine a "clean check" build takes 5.5s and a "clean opt" build
takes 16.7s, which is similar to some of the larger existing benchmarks,
though well short of the biggest three (`script-servo`, `style-servo`,
and `packed-simd`).