Check for msgpack before using it
When cdnjs.cloudflare.com is not allowed by noscript, the error handler
below would alert() the whole fetched data. Because that data is very
large, it can taken several seconds, before showing an alert. During
that time on Firefox for Android it is not possible to interact with
Firefox itself, so it is also not possible to open the noscript
settings to allow cdnjs.cloudflare.com.