Use async client with `wait()` for GH requests
Synchronous client causes panics due to Tokio's protection against using
a synchronous client inside an async context. It's not clear whether
this will work either; if not a larger refactor might become necessary.

We don't actually care whether the client here is blocking as we expect
a response very quickly (and GH is reliable enough, and important
enough, that it being down or slow to respond is already well-known to
everyone fairly quickly).