Move to async/await
This puts us on nightly, whereas we compiled on stable previously, but
due to the GitHub client no longer working synchronously (see d36697b34c89),
we need to use the async reqwest client. That's difficult to do well
without having async/await, since we need to thread the future(s)
through all of our functions, which without async/await is quite
painful. Since async/await is on track to stabilize soon, and no
significant changes are expected, this is deemed a worthwhile tradeoff.