Store just one "Stats" for each RunKind variant
We move the combining logic into a combine_with method on Stats. The
code which checked that the same statistics are collected for both Stats
is removed, since it seems not too critical that we check that -- if one
Stat has it and the other doesn't, that seems largely fine.