Only collect self-profile stats for the first run of a benchmark
This avoids the overheads of self-profile being reported in instructions
counts (and other stats).

We could also run without perf stat, but there's not much point in that.
Technically, this is interesting in that we're collecting the data from
perf stat twice now and collecting it together, but because we collate
that data via `min`, and `-Zself-profile` is basically guaranteed to be
slower/larger memory consumption/more faults, etc. , that should be fine
-- we still won't notice those overheads.