Prefer serializing self-profile data into a SoA
This both allows for much better compression (2 GB to 760 MB, with each timing
file going from 3.4 MB to 1.4 MB of compressed JSON), as well as far faster
deserialization (54 to 25 seconds). However, the new format is still fully
backwards compatible with the old (and we use this feature to deserialize data
from measureme in the collector).