Create interpolating iterator atop SeriesIterator
This interpolates in a step wise fashion.

Currently this commit doesn't work well with the summarization support embedded
into the series iterator, but in hindsight that support is fundamentally flawed:
the series iterator should only be iterating a single series, and a pass atop
multiple series iterators should aggregate those results.