Utilize aggregation of series chains instead of pre-interpolating
This doesn't yet remove the old interpolation code, but all use sites are
replaced.

A Series::All is "expanded" just before iterating over it to some set of crates,
and then each crate is converted to a SeriesIterator for use.

Aggregation also taints things that are interpolated going in to be interpolated
going out.