Refactor to use sqlite as primary db
rocksdb's limitations of opening only in a single process make it difficult to
deploy to ECS effectively, which we want to use. This commit refactors the
backend to use sqlite instead as the primary data store, which works well
locally (at current scales it would likely work well in production, too, though
eventually limitations on performance may make that problematic). It is planned
to add support for a postgres backend for production in the next few commits.