Make Connection take &self query methods
We want to be able to run queries with &self access to be able to run multiple
futures in "parallel" (but on one thread) them more effectively -- postgres
supports this natively, but sqlite in multi-threaded mode doesn't. So we
introduce a Mutex around the sqlite connection to force Sync implementation;
since we don't *actually* use the connections from multiple threads it should
not be a performance issue.