collector: passthrough env vars
This commit modifies the collector to stop clearing the environment when
invoking the fake rustc.

This is helpful for users of Nix development shells where environment
variables are required to inform tools of the location of C
dependencies, and for setting RUSTFLAGS to collect data from the same
build with different flags.

However, this change can result in some environment variables affecting
the data collection where that isn't intended.

Signed-off-by: David Wood <david@davidtw.co>