Incrementally make progress during collection
This permits restarting of the collection server without losing hour(s) of work.
It also removes the extremely long-lived (2h30m as of now) transaction in favor
of one wrapping each individual benchmark. Ideally, that would be by itself
"sufficiently small."

However, currently, even just limiting to one benchmark can mean a hour-long
transaction (for script-servo-2). This is a bit unfortunate, but at least
killing that transaction will be dealt with fairly smoothly.