Avoid collecting extraneous stuff on `eprintln` runs.
Currently, `collector eprintln` collects stderr output for the `cargo
rustc` command. This means it will collect rustc's stderr output for the
leaf crate, but *also* rustc's stderr output for build scripts that run
for the leaf crate, and *also* for Cargo's stderr output for all of the
above.

This means it is different to all the other profilers, which only
profile rustc on the leaf crate.

With this change, `collector eprintln` now only collects rustc's stderr
output for the leaf crate, so it matches the other profiler's behaviour.

The commit also changes `time-passes` to use the same collection
strategy as `eprintln`. This wasn't actually necessary, but I did it for
consistency.

Finally, the commit does some minor clean-ups.