Prepare build kinds in parallel
rustc is single threaded for a good portion of its compilation, and that means
that on many-core systems we're frequently only using one CPU core even during
preparation (where we don't care about noise). This builds all builds kinds in
parallel -- spawning N Cargos. We arrange for a jobserver to be passed to all
Cargos involved to make sure that we're not overloading the system with too many
processes.