Drop script-servo-2 benchmark
This benchmark takes 31% of our CI time, despite considerable time spent trying
to optimize it. However, we have not found that it is particularly notable when
compared to our other large crate benchmarks.

script-servo has also historically shown some amount of non-determinism in the
CGU partioning which also makes it an annoying benchmark.

Most people profiling locally just disable it because of how long it takes.