Graph/compare rustc bootstrap timing
The data is still pretty high variance, unfortunately, so we do not interleave
it with normal data on comparison pages (instead it is displayed afterwards).

Currently only wall-time for the bootstrap crates is collected, and it will be
displayed regardless of statistic on comparison pages. It's not clear that this
is a viable design indefinitely (we almost certainly want at least max-rss as
well); we may end up adding statistic column to rustc compilation stats as well.