bootstrap: at most one rustc running simultaneously
This commit adds another variance reduction strategy to rustc execution as part
of the bootstrap process, namely, rustbuild is configured to launch at most one
rustc (-j1) but then the individual rustc's are configured to have access to
full parallelism. This is not very realistic -- the build overall will likely be
slower than it could have been, when rustcs fight for resources -- but it should
hopefully still be representative of the total amount of work that needs to be
done for each rustc.

We care relatively speaking much more about being deterministic rather than
representative of "true" workloads, though, at least for now. It is relatively
unlikely that most improvements in build times will be major (>20%, which is
roughly the variance we observe today).