Do not compare crates that build in less than a second
Wall-time comparison for crates that build in less than a second are never going
to be reliable enough for us to make use of them. And, for the most part,
anything that builds in a second we don't actually care much about, since
it's almost certainly not a bottleneck in rustc build times (we have 100 second
long builds for rustc_middle, for example).