Select first before attempting an insert
As discovered in #773, Postgres increments the sequence even on aborted inserts.
This creates extremely large gaps between our ids, which is a problem (as we
then start overflowing the address space despite using a tiny amount of the
address space, e.g. 2.5k / 65k IDs for artifacts, requiring an upgrade to i32
artifact IDs).

sqlite does not seem to have this behavior, likely due to not supporting
concurrent write transactions (which is the only reason postgres defaults to the
eating behavior).