Fix raw downloads of self profile data
Upstream has migrated to single-file profile format which has been accounted for
in the collector (moving us over to uploading just the single compressed file
instead of a compressed tarball), but not yet in the server.

This fixes the raw download URLs; the rendered (crox + flamegraph) views will be
fixed in a separate commit.