Add diesel to the rustc perf test suite
As far as I know diesel is a rather strange workload for rustc.
According to some short measurements most of the time compiling diesel is
spend type checking the crate and resolving trait bounds. I see multiple
reasons for this:

* Diesel builts a complex abstract query dsl for SQL based on rust
generics. All fo this needs to be type checked.
* Diesel generates a ton of trait impls for tuples of various sizes.
There are features to set the supported max size to 16, 32, 64 and 128
tuple elements. As this is a benchmark, I've chosen to set it to 128 to
maximize the number of impls.

As consequence of this diesels compile times are quite sensitive to
changes touching the type system in general and the trait resolution in
detail. Any change that will introduce a behaviour which does not scale
well with the number of available trait impls will likely result in a huge
increase for this benchmark.