Add queued try commit count metric
We're currently tracking all commits in one metric, which reduces our ability to
tell what long queue lengths are caused by. Current presumed cause is try
builds, which are run in parallel on the submission side (try builds don't have
a bors queue), which makes them schedule simultaneously for us, increasing queue
length quickly. But this is not really easy to confirm.

We would also like to track a histogram (percentiles) of the latency before
getting results, but that will take more work.