Make sure to select from selected step
Not clear whether this will actually make a difference, but it seems not
impossible we were selecting from the "parent" collector_progress previously,
which would be wrong.

The intent is that the expected time is updated after a commit is fully
processed, not before; this ensures that we don't jump around as much on the
various print outs of this statistic.