Remove packed-simd benchmark
Upstream changes in rustc have made this no longer compile, and looking at the
README description (quoted below) and a quick perf profile did not indicate any
particularly interesting features that would make it worth continuing to
feature.

> Portable packed SIMD vectors. Exercises SIMD features heavily, and uses a compiler plugin.

This was added back in #268 but the reasons cited for its inclusion there don't
seem to actually reflect the current state of the benchmark (we never evaluated
test compilation time). I suspect that re-adding a more recent copy of stdsimd
may be relevant, but we indirectly measure compilation time there through the
self-bootstrap testing, so this is mostly duplicate work regardless.