Fix determination of whether or not to calculate variance
We were bailing if `variance_data.len() < MIN_PREVIOUS_COMMITS`, but
`variance_data.len()` isn't a commit count, but the count of test cases
for which we have at least one stat in the DB for the previous 50 to 100
commits. Bailing on this condition probably wasn't the intended behavior.

Instead, make the determination of whether or not to compute variance
for each test case individually. Only compute the variance for a test
case if there are at least `MIN_PREVIOUS_COMMITS` data points for it.