Sort patches by numeric index, not path
This may have caused patches to be applied in the wrong order; that should now
be fixed. In practice I think our patches likely worked out the same, but hard
to be sure. A quick check with sort and sort -n seems to imply things are
probably fine, but it may not have worked.