Update `cargo-llvm-lines` version and CI checks
An issue has been fixed in `cargo-llvm-lines` 0.4.12 which would cause
us to fail whenever profiling a crate which had any crate dependencies.
Update to 0.4.12 and add a CI check that profiles a crate with
dependencies to exercise this case.

The issue was that we would invoke a particular `cargo` executable and
`cargo-llvm-lines` would invoke a different one. This would cause us to
try to build the crate's dependencies a second time with the different
`cargo`. The `EXPECT_ONLY_WRAPPED_RUSTC` check would detect this case
and produce an error.