Upgrade to clap 3.
This requires moving away from the deprecated `clap_app!` macro. Clap's
derive API (formerly `StructOpt`) is nicer anyway, because it's not
stringly-typed.

There is quite a bit of furniture moving in `execute.rs` to better
separate the perf tools used for `bench_*` subcommands from those used
for the profile subcommands.