Expand logging for next_commit endpoint
We are seeing periodic failures to post results to pull requests, with the row
for the relevant build simply absent from the pull_request_build table in the
DB. Currently it seems non-obvious how this can happen: it seems like we should
always be enqueuing a build and subsequently attaching a commit in next_commit
for MissingReason::Master, which should always happen on at least one
next_commit invocation prior to us starting to benchmark.

This adds some extra logging that will hopefully help us debug this situation in
the future, as right now there's no obvious rationale for this happening.