Merge `diff_local` into `profile_local` (#1149)
The `diff_local` subcommand has a lot in common with `profile_local`,
but also has some unnecessary differences in behaviour, such as not
using an `<ID>` argument.

This commit removes `diff_local` and moves the diff functionality to
`profile_local` by adding a single new option to it, which makes things
more consistent and simplifies the code.

The commit also changes how ids are used by `profile_local` and
`bench_local`. Before, the `<ID>` argument was compulsory and it was
always used. Now it is optional. If not specified, the id used depends
on how rustc is specified:
- via a path: `Id`;
- via a `+toolchain`: the toolchain's name.

The commit also changes the naming of output files produced when doing
Cachegrind diffs.

The following examples show the interplay of these changes:

**OLD**

```
collector diff_local cachegrind \
    $RUSTC1 +rustc2 --builds Check --runs Full --include helloworld
```

raw output
- `cgout-before-helloworld-Check-Full`
- `cgout-rustc2-helloworld-Check-Full`

filtered raw output
- `cgfilt-before-helloworld-Check-Full`
- `cgfilt-rustc2-helloworld-Check-Full`

human-readable output
- `cgann-before-helloworld-Check-Full`
- `cgann-rustc2-helloworld-Check-Full`

filtered raw diff
- `cgdiff-before-rustc2-helloworld-Check-Full`

human-readable diff
- `cgann-before-rustc2-helloworld-Check-Full`
- `cgdiffann-latest`

**NEW**

```
collector profile_local cachegrind \
    $RUSTC1 --rustc2 +rustc2 --builds Check --runs Full --include helloworld
```

raw output
- `cgout-Id1-helloworld-Check-Full`
- `cgout-rustc2-helloworld-Check-Full`

filtered raw output
- `cgfilt-Id1-helloworld-Check-Full`
- `cgfilt-rustc2-helloworld-Check-Full`

human-readable output
- `cgann-Id1-helloworld-Check-Full`
- `cgann-rustc2-helloworld-Check-Full`

filtered raw diff
- `cgfilt-diff-Id1-rustc2-helloworld-Check-Full`

human-readable diff
- `cgann-diff-Id1-rustc2-helloworld-Check-Full`
- `cgann-diff-latest`

The diff files are named more consistently and the prefixes are more
consistent.

Also, with `diff_local` if you specified a toolchain via a `+abcdef`
revid it would use `abcdef` as the Id instead of `before` or `after`.
When diffing with `profile_local` it now uses `Id1`/`Id2`, for
consistency with the non-diff case.

Some of the "is this binary installed?" checks are moved around
so they are only run if necessary.

Finally, this commit documents the new behaviour in the README. (There
was no documentation at all for `diff_local`.) It also documents how
`+`-prefixed `<RUSTC>` options work, something that was also missing
from the docs.