Introduce the "stable" benchmark category.
(Also, temporarily, the "primary-and-stable" category. All the current
"primary-and-stable" benchmarks will become "stable" in the near
future.)

These categories are used in `perf-config.json` and the code that
chooses which benchmarks to execute. They allow the `supports_stable`
field to be removed.

However, when it comes to storing the category in the database, to avoid
database format changes, `Category` gets split into `supports_stable` (a
bool) and `category` (either "primary" or "secondary").

Some notable details:
- `Category` is moved to a different file, because it's no longer used
  by the DB code.
- Some methods are moved from `BenchmarkConfig` to `Category`.
- `Category` now derives `clap::ArgEnum`, which improves the `download`
  command's help message by listing the possible values.