Make `--include` and `--exclude` use exact matching.
In #1138 I changed `--include`/`--exclude` to do exact matching of
benchmark names. This was a good idea at the time, but since then we
have renamed many benchmarks. There is now no benchmark with a name that
is a prefix of another benchmark's name. Also, we now have version
numbers in many benchmark names, which are a pain to remember and type.

This commit lets you specify a benchmark name by prefix. E.g.
`--include=stm` will match `stm32f4-0.14.0`.