Be more rigorous and consistent about combining `Direction`s.
Currently `Direction`s are combined in two places.
- `populate_report`, which uses a match. It gives the primary and
  secondary directions equal weight.
- `summarize_run`, which uses `Option::or`. It gives the primary
  direction more weight; the secondary direction will only be used if
  the primary direction is `None`.

The latter seems suboptimal. If, for example, there are only
improvements in the primary benchmarks and only regressions in the
secondary benchmarks, the overall result should be "mixed results" not
"improvements".

This commit makes both places use the combination from
`populate_report`. `Direction` forms a simple lattice and that
combination is the "join" operation. `Option` is no longer needed.