Log preparation times
Getting these into the database would take more hassle and this is
enough for our logs to show the data. This will help us decide whether
to invest in moving preparation to be done in parallel across many
benchmarks or perhaps even in advance on a separate machine. Both of
those are quite a bit of refactoring, so we should figure out if they're
worth it first.