Remove the `keccak` compile-time benchmark.
It is prone to huge bimodal swings. It compiles down to a single
function that is massive because of extremely aggressive code
duplication via a declarative macro. It hugely stresses obligation
processing, which has arguably been over-optimized within rustc as a
result. `cranelift-codegen` provides similar (but less extreme) coverage
of obligation processing.

The latest version of the crate (`keccak-0.1.4`) is much less extreme,
with the largest function shrinking from 32,767 units to 1,680 units.