Generalize configuration of benchmarks
We will soon want to execute both compile and runtime benchmarks on multiple places. The current configuration of benchmarks was quite ad hoc. This PR introduces specialized config structs for both compile and runtime benchmarks (and also for configuration which is shared amongst them). It is a bit verbose, but also explicit.

Eventually, most of the call sites in `collector` should transition to using the `run_benchmarks` function, but so far runtime benchmarks only run for `bench_runtime_local` and `bench_published`.