Eagerly load self-profile files into memory
This means that we can avoid generating a unique directory for the self-profiles (reverted in the last commit), and we also avoid loading the self-profile files twice from disk. The disadvantage is that all self-profiles for a single benchmark (so for all combinations of profiles and scenarios) is now held in memory. But I think that shouldn't be an issue.