Add function for getting the last N completed requests
The function returns errors associated with the requests, including both the error benchmark/context name and the error content. The query is also prepared, because it should be as fast as possible to not block the status page.

Also adds some basic test builders for easier preparation of test state in tests.